third_party/WebKit/LayoutTests/http/tests/security/referrer-policy-worker-no-referrer.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>
    async_test(function () {
        var worker = new Worker('http://127.0.0.1:8000/security/contentSecurityPolicy/resources/worker.php?type=report-referrer&referrerpolicy=' +
                                encodeURIComponent('no-referrer'));
        worker.onmessage = this.step_func(function (event) {
            assert_equals("HTTP Referer header is empty", event.data);
            this.done();
        });
    }, "Worker with no-referrer CSP");
</script>
</body>
</html>
